refactor : 패키지 구조 이동 및 primary 어노테이션 추가
